+++ src/commands/import_command.cpp
@@ -3,14 +3,16 @@
 #include <plog/Log.h>
 
 #include "importer/radio_sure_importer.hpp"
+#include "importer/radio_browser_importer.hpp"
 
 #include <memory>
 
 ImportCommand::ImportCommand() : Command("import", "Import radio stations from different sources"), m_importerByName() {
         auto radioSureImporter = std::unique_ptr<RadioSureImporter>(new RadioSureImporter());
         m_importerByName.insert(std::make_pair(radioSureImporter->getName(), std::move(radioSureImporter)));
+        auto radioBrowserImporter = std::unique_ptr<RadioBrowserImporter>(new RadioBrowserImporter());
+        m_importerByName.insert(std::make_pair(radioBrowserImporter->getName(), std::move(radioBrowserImporter)));
         m_cli.addOption('i', "input", true, "Input to import stations from. Depends on the type an could be a file or URL.");
-        // TODO add real names of the importer
         m_cli.addOption('t', "type", true, "Type of imported data. Supported types are: radio-sure (requires a file as input).");
 }
 
@@ -24,19 +26,16 @@ void ImportCommand::execute(const std::vector<std::string>& args) {
         return;
     }
 
-    if (!m_cli.hasValue('t') || !m_cli.hasValue('i')) {
-        LOG(plog::warning) << "Missing arguments. Try " << m_name << " --help for more information.";
-        return;
-    }
-
-    auto importer = m_importerByName.find(m_cli.getValue('t'));
+    // TODO Name should be static, so it must not be hard coded here
+    auto importer = m_importerByName.find(m_cli.getValue('t', "radio-browser"));
 
     if (importer == m_importerByName.end()) {
         LOG(plog::error) << "Unknown importer type " << m_cli.getValue('t') << ". Try " << m_name << " --help for more information.";
     } else {
         auto stationsDao = createStationsDao();
         stationsDao->open(m_cli.getValue('f', getDefaultFile()));
-        LOG(plog::info) << "importing from " << m_cli.getValue('i') << " using " << importer->second->getName() << " importer";
-        importer->second->import(m_cli.getValue('i'), stationsDao);
+        const std::string input = m_cli.getValue('i', "");
+        LOG(plog::info) << "importing from " << input << " using " << importer->second->getName() << " importer";
+        importer->second->import(input, stationsDao);
     }
 }